backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["idabblewith <jprincedev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.14,>=3.12.11"
djangorestframework = "^3.16.0"
sentry-sdk = { extras = ["django"], version = "^2.25.1" }
whitenoise = { extras = ["brotli"], version = "^6.9.0" }
django-environ = "^0.12.0"
requests = "^2.32.3"
django-cors-headers = "^4.7.0"
psycopg = "^3.2.6"
pyjwt = "^2.10.1"
pillow = "^11.2.1"
django-storages = { extras = ["azure"], version = "^1.14.6" }
gunicorn = "^23.0.0"
psycopg-binary = "^3.2.9"
dj-database-url = "^3.0.0"
dbca-utils = "^2.1.3"
black = "^25.1.0"
djangorestframework-simplejwt = "^5.5.1"
matplotlib = "^3.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
